Add new test features and removed the  redundancy between deepmd and meam

+++ dpgen/auto_test/gen_01_eos.py
@@ -91,13 +91,23 @@ def make_vasp(jdata, conf_dir) :
         # print(vol_path, vasp.poscar_vol('POSCAR') / vasp.poscar_natoms('POSCAR'))
         os.chdir(cwd)
 
-def make_deepmd_lammps (jdata, conf_dir) :
-    deepmd_model_dir = jdata['deepmd_model_dir']
-    deepmd_type_map = jdata['deepmd_type_map']
-    ntypes = len(deepmd_type_map)    
-    deepmd_model_dir = os.path.abspath(deepmd_model_dir)
-    deepmd_models = glob.glob(os.path.join(deepmd_model_dir, '*pb'))
-    deepmd_models_name = [os.path.basename(ii) for ii in deepmd_models]
+def make_lammps (jdata, conf_dir,task_type) :
+    fp_params = jdata['lammps_params']
+    model_dir = fp_params['model_dir']
+    type_map = fp_params['type_map'] 
+    model_dir = os.path.abspath(model_dir)
+    model_name =fp_params['model_name']
+    if not model_name and task_type =='deepmd':
+        models = glob.glob(os.path.join(model_dir, '*pb'))
+        model_name = [os.path.basename(ii) for ii in models]
+    else:
+        models = [os.path.join(model_dir,ii) for ii in model_name]
+
+    model_param = {'model_name' :      fp_params['model_name'],
+                  'param_type':          fp_params['model_param_type']}
+    
+    ntypes = len(type_map)
+
     vol_start = jdata['vol_start']
     vol_end = jdata['vol_end']
     vol_step = jdata['vol_step']
@@ -129,20 +139,23 @@ def make_deepmd_lammps (jdata, conf_dir) :
     # structrure
     ss = Structure.from_file(to_poscar)
     # lmp path
-    lmp_path = os.path.join(task_path, 'deepmd')
+    lmp_path = os.path.join(task_path, task_type)
     os.makedirs(lmp_path, exist_ok = True)
     # # lmp conf
     # conf_file = os.path.join(lmp_path, 'conf.lmp')
     # lammps.cvt_lammps_conf(to_poscar, conf_file)
     # ptypes = vasp.get_poscar_types(to_poscar)
     # lammps.apply_type_map(conf_file, deepmd_type_map, ptypes)
-    os.chdir(lmp_path)
-    for ii in deepmd_models_name :
-        if os.path.exists(ii) :
-            os.remove(ii)
-    for (ii,jj) in zip(deepmd_models, deepmd_models_name) :
+    if task_type =='deepmd':
+        os.chdir(lmp_path)
+        for ii in model_name :
+            if os.path.exists(ii) :
+                os.remove(ii)
+        for (ii,jj) in zip(models, model_name) :
             os.symlink(os.path.relpath(ii), jj)
-    share_models = glob.glob(os.path.join(lmp_path, '*pb'))
+        share_models = glob.glob(os.path.join(lmp_path, '*pb'))
+    else:
+        share_models =models
 
     for vol in np.arange(vol_start, vol_end, vol_step) :
         vol_path = os.path.join(lmp_path, 'vol-%.2f' % vol)        
@@ -150,7 +163,7 @@ def make_deepmd_lammps (jdata, conf_dir) :
         os.makedirs(vol_path, exist_ok = True)
         os.chdir(vol_path)
         #print(vol_path)
-        for ii in ['conf.lmp', 'conf.lmp'] + deepmd_models_name :
+        for ii in ['conf.lmp', 'conf.lmp'] + model_name :
             if os.path.exists(ii) :
                 os.remove(ii)                
         # # link conf
@@ -161,99 +174,37 @@ def make_deepmd_lammps (jdata, conf_dir) :
         scale_ss.to('POSCAR', 'POSCAR')
         lammps.cvt_lammps_conf('POSCAR', 'conf.lmp')
         ptypes = vasp.get_poscar_types('POSCAR')
-        lammps.apply_type_map('conf.lmp', deepmd_type_map, ptypes)
+        lammps.apply_type_map('conf.lmp', type_map, ptypes)
         # link models
-        for (ii,jj) in zip(share_models, deepmd_models_name) :
+        for (ii,jj) in zip(share_models, model_name) :
             os.symlink(os.path.relpath(ii), jj)
         # make lammps input
         scale = (vol / vpa) ** (1./3.)
-        fc = lammps.make_lammps_press_relax('conf.lmp', ntypes, scale,lammps.inter_deepmd, deepmd_models_name)
+        if task_type=='deepmd':
+            fc = lammps.make_lammps_press_relax('conf.lmp', ntypes, scale,lammps.inter_deepmd, model_name)
+        elif task_type =='meam':
+            fc = lammps.make_lammps_press_relax('conf.lmp', ntypes, scale,lammps.inter_meam, model_param)    
         with open(os.path.join(vol_path, 'lammps.in'), 'w') as fp :
             fp.write(fc)
         os.chdir(cwd)
 
-def make_deepmd_lammps_fixv (jdata, conf_dir) :
-    deepmd_model_dir = jdata['deepmd_model_dir']
-    deepmd_type_map = jdata['deepmd_type_map']
-    ntypes = len(deepmd_type_map)    
-    deepmd_model_dir = os.path.abspath(deepmd_model_dir)
-    deepmd_models = glob.glob(os.path.join(deepmd_model_dir, '*pb'))
-    deepmd_models_name = [os.path.basename(ii) for ii in deepmd_models]
-    vol_start = jdata['vol_start']
-    vol_end = jdata['vol_end']
-    vol_step = jdata['vol_step']
+def make_lammps_fixv (jdata, conf_dir,task_type) :
+    fp_params = jdata['lammps_params']
+    model_dir = fp_params['model_dir']
+    type_map = fp_params['type_map'] 
+    model_dir = os.path.abspath(model_dir)
+    model_name =fp_params['model_name']
+    if not model_name and task_type =='deepmd':
+        models = glob.glob(os.path.join(model_dir, '*pb'))
+        model_name = [os.path.basename(ii) for ii in models]
+    else:
+        models = [os.path.join(model_dir,ii) for ii in model_name]
 
-    # get equi props
-    equi_path = re.sub('confs', global_equi_name, conf_dir)
-    task_path = re.sub('confs', global_task_name, conf_dir)
-    equi_path = os.path.join(equi_path, 'deepmd')
-    task_path = os.path.join(task_path, 'deepmd')
-    equi_path = os.path.abspath(equi_path)
-    task_path = os.path.abspath(task_path)
-    equi_log = os.path.join(equi_path, 'log.lammps')
-    equi_dump = os.path.join(equi_path, 'dump.relax')
-    os.makedirs(task_path, exist_ok = True)
-    task_poscar = os.path.join(task_path, 'POSCAR')
-    lammps.poscar_from_last_dump(equi_dump, task_poscar, deepmd_type_map)
-
-    cwd = os.getcwd()
-    volume = vasp.poscar_vol(task_poscar)
-    natoms = vasp.poscar_natoms(task_poscar)
-    vpa = volume / natoms
-    # structrure
-    ss = Structure.from_file(task_poscar)
-    # make lammps.in
-    fc = lammps.make_lammps_equi('conf.lmp', 
-                                 ntypes, 
-                                 lammps.inter_deepmd, 
-                                 deepmd_models_name, 
-                                 change_box = False)
-    f_lammps_in = os.path.join(task_path, 'lammps.in')
-    with open(f_lammps_in, 'w') as fp :
-        fp.write(fc)
-        
-    os.chdir(task_path)
-    for ii in deepmd_models_name :
-        if os.path.exists(ii) :
-            os.remove(ii)
-    for (ii,jj) in zip(deepmd_models, deepmd_models_name) :
-            os.symlink(os.path.relpath(ii), jj)
-    share_models = glob.glob(os.path.join(task_path, '*pb'))
-
-    # make vols
-    for vol in np.arange(vol_start, vol_end, vol_step) :
-        vol_path = os.path.join(task_path, 'vol-%.2f' % vol)        
-        print('# generate %s' % (vol_path))
-        os.makedirs(vol_path, exist_ok = True)
-        os.chdir(vol_path)
-        for ii in ['conf.lmp', 'conf.lmp', 'lammps.in'] + deepmd_models_name :
-            if os.path.exists(ii) :
-                os.remove(ii)                
-        # make conf
-        scale_ss = ss.copy()
-        scale_ss.scale_lattice(vol * natoms)
-        scale_ss.to('POSCAR', 'POSCAR')
-        lammps.cvt_lammps_conf('POSCAR', 'conf.lmp')
-        ptypes = vasp.get_poscar_types('POSCAR')
-        lammps.apply_type_map('conf.lmp', deepmd_type_map, ptypes)
-        # link lammps.in
-        os.symlink(os.path.relpath(f_lammps_in), 'lammps.in')
-        # link models
-        for (ii,jj) in zip(share_models, deepmd_models_name) :
-            os.symlink(os.path.relpath(ii), jj)
-        # make lammps input
-        os.chdir(cwd)
-
-def make_meam_lammps_fixv (jdata, conf_dir) :
-    meam_potfile_dir = jdata['meam_potfile_dir']
-    meam_potfile_dir = os.path.abspath(meam_potfile_dir)
-    meam_potfile = jdata['meam_potfile']
-    meam_potfile = [os.path.join(meam_potfile_dir,ii) for ii in meam_potfile]
-    meam_potfile_name = jdata['meam_potfile']
-    type_map = jdata['meam_type_map']
+    model_param = {'model_name' :      fp_params['model_name'],
+                  'param_type':          fp_params['model_param_type']}
+    
     ntypes = len(type_map)
-    meam_param = {'meam_potfile' :      jdata['meam_potfile'],
-                  'meam_type':          jdata['meam_param_type']}
+
 
     vol_start = jdata['vol_start']
     vol_end = jdata['vol_end']
@@ -262,11 +213,10 @@ def make_meam_lammps_fixv (jdata, conf_dir) :
     # get equi props
     equi_path = re.sub('confs', global_equi_name, conf_dir)
     task_path = re.sub('confs', global_task_name, conf_dir)
-    equi_path = os.path.join(equi_path, 'meam')
-    task_path = os.path.join(task_path, 'meam')
+    equi_path = os.path.join(equi_path, task_type)
+    task_path = os.path.join(task_path, task_type)
     equi_path = os.path.abspath(equi_path)
     task_path = os.path.abspath(task_path)
-    equi_log = os.path.join(equi_path, 'log.lammps')
     equi_dump = os.path.join(equi_path, 'dump.relax')
     os.makedirs(task_path, exist_ok = True)
     task_poscar = os.path.join(task_path, 'POSCAR')
@@ -279,21 +229,41 @@ def make_meam_lammps_fixv (jdata, conf_dir) :
     # structrure
     ss = Structure.from_file(task_poscar)
     # make lammps.in
-    fc = lammps.make_lammps_equi('conf.lmp', 
+    if task_type=='deepmd':
+        fc = lammps.make_lammps_equi('conf.lmp', 
+                                 ntypes, 
+                                 lammps.inter_deepmd, 
+                                 model_name, 
+                                 change_box = False)
+    elif task_type=='meam':
+        fc = lammps.make_lammps_equi('conf.lmp', 
                                  ntypes, 
                                  lammps.inter_meam, 
-                                 meam_param, 
+                                 model_param, 
                                  change_box = False)
     f_lammps_in = os.path.join(task_path, 'lammps.in')
     with open(f_lammps_in, 'w') as fp :
         fp.write(fc)
+
+    if task_type =='deepmd':    
+        os.chdir(task_path)
+        for ii in model_name :
+            if os.path.exists(ii) :
+                os.remove(ii)
+        for (ii,jj) in zip(models, model_name) :
+            os.symlink(os.path.relpath(ii), jj)
+        share_models = glob.glob(os.path.join(task_path, '*pb'))
+    else:
+        share_models = models
+
+
     # make vols
     for vol in np.arange(vol_start, vol_end, vol_step) :
         vol_path = os.path.join(task_path, 'vol-%.2f' % vol)        
         print('# generate %s' % (vol_path))
         os.makedirs(vol_path, exist_ok = True)
         os.chdir(vol_path)
-        for ii in ['conf.lmp', 'conf.lmp', 'lammps.in'] + meam_potfile_name :
+        for ii in ['conf.lmp', 'conf.lmp', 'lammps.in'] + model_name :
             if os.path.exists(ii) :
                 os.remove(ii)                
         # make conf
@@ -306,10 +276,11 @@ def make_meam_lammps_fixv (jdata, conf_dir) :
         # link lammps.in
         os.symlink(os.path.relpath(f_lammps_in), 'lammps.in')
         # link models
-        for (ii,jj) in zip(meam_potfile, meam_potfile_name) :
+        for (ii,jj) in zip(share_models,model_name) :
             os.symlink(os.path.relpath(ii), jj)
         # make lammps input
         os.chdir(cwd)
+
 
 def _main() :
     parser = argparse.ArgumentParser(
@@ -330,16 +301,11 @@ def _main() :
     # print('generate %s task with conf %s' % (args.TASK, args.CONF))
     if args.TASK == 'vasp':
         make_vasp(jdata, args.CONF)               
-    elif args.TASK == 'deepmd' :
+    elif args.TASK == 'deepmd' or args.TASK =='meam':
         if args.fix_shape :
-            make_deepmd_lammps_fixv(jdata, args.CONF)
+            make_lammps_fixv(jdata, args.CONF,args.TASK)
         else :
-            make_deepmd_lammps(jdata, args.CONF)        
-    elif args.TASK == 'meam' :
-        if args.fix_shape:
-            make_meam_lammps_fixv(jdata, args.CONF)
-        else :
-            raise RuntimeError("not implemented ", args.TASK)            
+            make_lammps(jdata, args.CONF,args.TASK)              
     else :
         raise RuntimeError("unknow task ", args.TASK)
